core: refactor + (twofactor, verify)
