--- backend/src/api/sessions.rs
@@ -1,190 +1,0 @@
-use super::{
-    microservices::{
-        database::core::insert_user,
-        redis::{increment_lock_key, insert_id, insert_session},
-    },
-    models::{Action, RedisAccount, RedisAction, WebsitePath},
-    twofactor::spawn_code_task,
-};
-use crate::{AppError, AppState};
-use axum::http::{
-    HeaderValue,
-    header::{HeaderMap, SET_COOKIE},
-};
-use axum_extra::extract::CookieJar;
-use cookie::{Cookie, CookieJar as cookieCookieJar, SameSite::Strict, time::Duration};
-use once_cell::sync::Lazy;
-use std::sync::Arc;
-use uuid::Uuid;
-
-static COOKIES_TO_CLEAR: Lazy<Vec<&'static str>> = Lazy::new(|| {
-    vec![
-        RedisAction::Session.as_ref(),
-        RedisAction::Forgot.as_ref(),
-        RedisAction::Update.as_ref(),
-        RedisAction::Auth.as_ref(),
-    ]
-});
-
-pub static CLEARED_COOKIES_SWAP: Lazy<cookieCookieJar> =
-    Lazy::new(|| cleared_cookies_for(WebsitePath::BoilerSwap));
-
-pub fn cleared_cookies_for(website_path: WebsitePath) -> cookieCookieJar {
-    let mut jar = cookieCookieJar::new();
-
-    for &old_cookie in COOKIES_TO_CLEAR.iter() {
-        let expired = Cookie::build(old_cookie)
-            .path(format!("/{}", website_path.as_ref()))
-            .http_only(true)
-            .secure(true)
-            .same_site(Strict)
-            .max_age(Duration::seconds(0));
-
-        jar.add(expired);
-    }
-
-    jar
-}
-
-pub fn generate_cookie(
-    key: &str,
-    value: &str,
-    ttl: i64,
-    website_path: WebsitePath,
-) -> Result<HeaderMap, AppError> {
-    let mut jar = match website_path {
-        WebsitePath::BoilerSwap => CLEARED_COOKIES_SWAP.clone(),
-        WebsitePath::Photos => {
-            return Err(AppError::Config(
-                "Invalid path for cookie generation".into(),
-            ));
-        }
-        WebsitePath::Home => {
-            return Err(AppError::Config(
-                "Invalid path for cookie generation".into(),
-            ));
-        }
-    };
-
-    let new_cookie = Cookie::build((key.to_owned(), value.to_owned()))
-        .path(format!("/{}", website_path.as_ref()))
-        .http_only(true)
-        .secure(true)
-        .same_site(Strict)
-        .max_age(Duration::seconds(ttl));
-
-    jar.add(new_cookie);
-
-    let mut headers = HeaderMap::new();
-
-    for cookie in jar.delta() {
-        headers.append(
-            SET_COOKIE,
-            HeaderValue::from_str(&cookie.to_string()).unwrap(),
-        );
-    }
-
-    Ok(headers)
-}
-
-pub fn get_cookie(headers: &HeaderMap, key: &str) -> Option<String> {
-    CookieJar::from_headers(headers)
-        .get(key)
-        .map(|cookie| cookie.value().to_string())
-}
-
-pub async fn create_temporary_session(
-    state: Arc<AppState>,
-    result: &Option<String>,
-    redis_account: &RedisAccount,
-    redis_action: RedisAction,
-    forgot_key: &Option<String>,
-    code_key: &Option<String>,
-    website_path: WebsitePath,
-) -> Result<HeaderMap, AppError> {
-    if redis_action != RedisAction::Update {
-        spawn_code_task(
-            state.clone(),
-            redis_account.email.clone(),
-            redis_account.code.clone(),
-            forgot_key.clone(),
-            website_path.as_ref().to_string(),
-        );
-
-        increment_lock_key(
-            state.clone(),
-            website_path.as_ref(),
-            code_key.as_ref().unwrap(),
-            &redis_account.email.clone(),
-            &state.config.authentication.max_codes_duration_seconds,
-            &state.config.authentication.max_codes,
-        )
-        .await?;
-    }
-
-    let serialized = match result {
-        Some(result) => result,
-        None => &serde_json::to_string(&redis_account)?,
-    };
-
-    let id = Uuid::new_v4().to_string();
-
-    insert_id(
-        state.clone(),
-        &format!(
-            "{}:{}:{}",
-            website_path.as_ref(),
-            redis_action.as_ref(),
-            &id
-        ),
-        serialized,
-        state
-            .config
-            .session
-            .temporary_session_duration_seconds
-            .into(),
-    )
-    .await?;
-
-    generate_cookie(
-        redis_action.as_ref(),
-        &id,
-        state
-            .config
-            .session
-            .temporary_session_duration_seconds
-            .into(),
-        website_path,
-    )
-}
-
-pub async fn create_session(
-    state: Arc<AppState>,
-    redis_account: &RedisAccount,
-    redis_action: RedisAction,
-    redis_action_secondary: RedisAction,
-    website_path: WebsitePath,
-) -> Result<HeaderMap, AppError> {
-    if redis_account.action == Action::Signup {
-        insert_user(state.clone(), redis_account.clone()).await?;
-    }
-
-    let session_id = Uuid::new_v4().to_string();
-
-    insert_session(
-        state.clone(),
-        website_path.as_ref(),
-        redis_action.as_ref(),
-        &session_id,
-        redis_action_secondary.as_ref(),
-        &redis_account.email,
-    )
-    .await?;
-
-    generate_cookie(
-        redis_action.as_ref(),
-        &session_id,
-        state.config.session.session_duration_seconds.into(),
-        website_path,
-    )
-}
