+++ backend/src/api/web/mod.rs
@@ -1,0 +1,8 @@
+pub mod cookies;
+pub mod handlers;
+pub mod home;
+pub mod lock;
+pub mod sessions;
+pub mod swap;
+pub mod twofactor;
+pub mod verify;
